Purpose: To reduce player difficulty in finding targetable planets for the operations that use the Min Planet % Owned op setting this command was added to allow the player to select AttackerHouse DefenderHouse and Planet% and they will receive a simple list of planet hyperlinks.
Added server.campaign.commands.FindContestedPlanets as the command.
In server.campaign.CampaignMain - added the findcp command to the commands table.
In client.gui.CMainFrame - added Campaign menu>Attack>Find Contested Planets item and code to interact with user and send the command to server.

all changes are marked with "BarukKhazad 20151129"